+++ yazi-binding/src/config/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(fetcher fetchers preloader preloaders previewer previewers spotter spotters);
+yazi_macro::mod_flat!(fetcher fetchers open_rule open_rules opener opener_rule opener_rules preloader preloaders previewer previewers spotter spotters);