+++ yazi-codegen/src/lib.rs
@@ -1,6 +1,9 @@
 use proc_macro::TokenStream;
-use quote::{format_ident, quote};
-use syn::{Data, DeriveInput, Fields, parse_macro_input};
+use quote::quote;
+mod helper;
+use syn::{DeriveInput, parse_macro_input};
+
+use crate::helper::{generics_with_de, has_serde_attr, ident_name, named_fields};
 
 #[proc_macro_derive(DeserializeOver)]
 pub fn deserialize_over(input: TokenStream) -> TokenStream {
@@ -18,47 +21,48 @@ pub fn deserialize_over1(input: TokenStream) -> TokenStream {
 	let DeriveInput { ident, generics, data, .. } = parse_macro_input!(input as DeriveInput);
 	let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
-	let assignments = match data {
-		Data::Struct(struct_) => match struct_.fields {
-			Fields::Named(fields) => {
-				let mut assignments = Vec::with_capacity(fields.named.len());
+	let visitor_generics = generics_with_de(&generics);
+	let (impl_visitor_generics, ..) = visitor_generics.split_for_impl();
 
-				for field in fields.named {
-					let field_ident = &field.ident;
-					let field_name = field_ident.as_ref().unwrap().to_string();
+	let fields: Vec<_> = named_fields(data).into_iter().map(|f| f.ident.unwrap()).collect();
+	let match_arms = fields.iter().map(|f| {
+		let name = ident_name(f);
+		quote! { #name => self.0.#f = map.next_value_seed(DeserializeOverSeed(self.0.#f))?, }
+	});
 
-					assignments.push(quote! {
-						if let Some(value) = table.remove(#field_name) {
-							if !matches!(value.get_ref(), toml::de::DeValue::Table(_)) {
-								_ = deserialize_spanned::<toml::Table>(value)?;
-								return Err(serde::de::Error::custom(format!("expected top-level `{}` to be a TOML table", #field_name)));
-							}
-
-							let span = value.span();
-							if let toml::de::DeValue::Table(table) = value.into_inner() {
-								self.#field_ident = self.#field_ident.deserialize_over_with(toml::Spanned::new(span, table))?;
-							}
-						}
-						self.#field_ident = self.#field_ident.deserialize_over_hook()?;
-					});
-				}
-
-				assignments
-			}
-			_ => panic!("DeserializeOver1 only supports structs with named fields"),
-		},
-		_ => panic!("DeserializeOver1 only supports structs"),
-	};
+	let hook_fields = fields.iter().map(|f| {
+		quote! { #f: deserialized.#f.deserialize_over_hook().map_err(Error::custom)? }
+	});
 
 	quote! {
 		impl #impl_generics yazi_shim::toml::DeserializeOverWith for #ident #ty_generics #where_clause {
-			fn deserialize_over_with<'de>(mut self, table: toml::Spanned<toml::de::DeTable<'de>>) -> Result<Self, toml::de::Error> {
-				use yazi_shim::toml::{DeserializeOverHook, DeserializeOverWith, deserialize_spanned};
+			fn deserialize_over_with<'__de, __D: serde::Deserializer<'__de>>(self, de: __D) -> Result<Self, __D::Error> {
+				use std::borrow::Cow;
+				use serde::de::{Error, IgnoredAny, MapAccess, Visitor};
+				use yazi_shim::toml::{DeserializeOverHook, DeserializeOverSeed};
 
-				let mut table = table.into_inner();
-				#(#assignments)*
+				struct V #impl_generics (#ident #ty_generics) #where_clause;
 
-				self.deserialize_over_hook()
+				impl #impl_visitor_generics Visitor<'__de> for V #ty_generics #where_clause {
+					type Value = #ident #ty_generics;
+
+					fn expecting(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+						f.write_str("a map")
+					}
+
+					fn visit_map<__M: MapAccess<'__de>>(mut self, mut map: __M) -> Result<Self::Value, __M::Error> {
+						while let Some(key) = map.next_key::<Cow<str>>()? {
+							match key.as_ref() {
+								#(#match_arms)*
+								_ => { map.next_value::<IgnoredAny>()?; }
+							}
+						}
+						Ok(self.0)
+					}
+				}
+
+				let deserialized = de.deserialize_map(V(self))?;
+				Ok(Self { #(#hook_fields,)* })
 			}
 		}
 	}
@@ -70,47 +74,79 @@ pub fn deserialize_over2(input: TokenStream) -> TokenStream {
 	let DeriveInput { ident, generics, data, .. } = parse_macro_input!(input as DeriveInput);
 	let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
-	let assignments = match data {
-		Data::Struct(struct_) => match struct_.fields {
-			Fields::Named(fields) => {
-				let mut assignments = Vec::with_capacity(fields.named.len());
+	let visitor_generics = generics_with_de(&generics);
+	let (impl_visitor_generics, ..) = visitor_generics.split_for_impl();
 
-				for (index, field) in fields.named.into_iter().enumerate() {
-					let field_ident = field.ident;
-					let field_name = field_ident.as_ref().unwrap().to_string();
-					let field_ty = field.ty;
-					let field_attrs = field.attrs;
-					let field_helper = format_ident!("__DeserializeOver2Field{index}");
+	let (mut match_arms, mut post_fields) = (vec![], vec![]);
+	for field in named_fields(data) {
+		let (field_ident, field_ty) = (field.ident, field.ty);
+		let field_name = ident_name(field_ident.as_ref().unwrap());
 
-					assignments.push(quote! {
-						#[derive(serde::Deserialize)]
-						struct #field_helper #impl_generics (
-							#(#field_attrs)*
-							#field_ty
-						) #where_clause;
+		if has_serde_attr(&field.attrs, "skip") {
+			continue;
+		}
 
-						if let Some(value) = table.remove(#field_name) {
-							self.#field_ident = deserialize_spanned::<#field_helper #ty_generics>(value)?.0;
-						}
-					});
+		if field_name == "selector" && has_serde_attr(&field.attrs, "flatten") {
+			match_arms.push(quote! {
+				"url" => selector_url = Some(map.next_value()?),
+				"mime" => selector_mime = Some(map.next_value()?),
+			});
+			post_fields.push(quote! {
+				self.0.#field_ident = Selector::new(
+					selector_url.or(self.0.#field_ident.url),
+					selector_mime.or(self.0.#field_ident.mime)
+				).map_err(Error::custom)?;
+			});
+			continue;
+		}
+
+		let serde_attrs: Vec<_> = field.attrs.iter().filter(|a| a.path().is_ident("serde")).collect();
+		if serde_attrs.is_empty() {
+			match_arms.push(quote! { #field_name => self.0.#field_ident = map.next_value()?, });
+		} else {
+			match_arms.push(quote! {
+				#field_name => {
+					#[derive(serde::Deserialize)]
+					struct H #impl_generics(#(#serde_attrs)* #field_ty,) #where_clause;
+					self.0.#field_ident = map.next_value::<H #ty_generics>()?.0;
 				}
-
-				assignments
-			}
-			_ => panic!("DeserializeOver2 only supports structs with named fields"),
-		},
-		_ => panic!("DeserializeOver2 only supports structs"),
-	};
+			});
+		}
+	}
 
 	quote! {
 		impl #impl_generics yazi_shim::toml::DeserializeOverWith for #ident #ty_generics #where_clause {
-			fn deserialize_over_with<'de>(mut self, table: toml::Spanned<toml::de::DeTable<'de>>) -> Result<Self, toml::de::Error> {
-				use yazi_shim::toml::deserialize_spanned;
+			fn deserialize_over_with<'__de, __D: serde::Deserializer<'__de>>(self, de: __D) -> Result<Self, __D::Error> {
+				use std::borrow::Cow;
+				use serde::de::{Error, IgnoredAny, MapAccess, Visitor};
+				use crate::Selector;
 
-				let mut table = table.into_inner();
-				#(#assignments)*
+				struct V #impl_generics (#ident #ty_generics) #where_clause;
 
-				Ok(self)
+				impl #impl_visitor_generics Visitor<'__de> for V #ty_generics #where_clause {
+					type Value = #ident #ty_generics;
+
+					fn expecting(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+						f.write_str("a map")
+					}
+
+					fn visit_map<__M: MapAccess<'__de>>(mut self, mut map: __M) -> Result<Self::Value, __M::Error> {
+						let mut selector_url: Option<crate::Pattern> = None;
+						let mut selector_mime: Option<crate::Pattern> = None;
+
+						while let Some(key) = map.next_key::<Cow<str>>()? {
+							match key.as_ref() {
+								#(#match_arms)*
+								_ => { map.next_value::<IgnoredAny>()?; }
+							}
+						}
+
+						#(#post_fields)*
+						Ok(self.0)
+					}
+				}
+
+				de.deserialize_map(V(self))
 			}
 		}
 	}