+++ yazi-config/src/plugin/spotter.rs
@@ -1,9 +1,10 @@
-use std::ops::Deref;
+use std::{borrow::Cow, ops::Deref, sync::Arc};
 
 use serde::Deserialize;
+use yazi_fs::File;
 use yazi_shared::{Id, event::Action};
 
-use crate::{Mixable, Pattern, Selectable, Selector, plugin::spotter_id};
+use crate::{Mixable, Pattern, Selectable, Selector, plugin::{Spotters, spotter_id}};
 
 #[derive(Debug, Deserialize)]
 pub struct Spotter {
@@ -31,4 +32,47 @@ impl Mixable for Spotter {
 	fn any_file(&self) -> bool { self.selector.any_file() }
 
 	fn any_dir(&self) -> bool { self.selector.any_dir() }
+}
+
+// --- Matcher
+#[derive(Default)]
+pub struct SpotterMatcher<'a> {
+	pub spotters: Arc<Vec<Arc<Spotter>>>,
+	pub id:       Id,
+	pub file:     Option<Cow<'a, File>>,
+	pub mime:     Option<Cow<'a, str>>,
+	pub all:      bool,
+	pub offset:   usize,
+}
+
+impl From<&Spotters> for SpotterMatcher<'_> {
+	fn from(spotters: &Spotters) -> Self {
+		Self { spotters: spotters.load_full(), all: true, ..Default::default() }
+	}
+}
+
+impl SpotterMatcher<'_> {
+	pub fn matches(&self, spotter: &Spotter) -> bool {
+		if self.all {
+			true
+		} else if self.id != Id::ZERO {
+			spotter.id == self.id
+		} else {
+			spotter.match_with(self.file.as_deref(), self.mime.as_deref())
+		}
+	}
+}
+
+impl Iterator for SpotterMatcher<'_> {
+	type Item = Arc<Spotter>;
+
+	fn next(&mut self) -> Option<Self::Item> {
+		while let Some(spotter) = self.spotters.get(self.offset) {
+			self.offset += 1;
+			if self.matches(spotter) {
+				return Some(spotter.clone());
+			}
+		}
+		None
+	}
 }