+++ yazi-config/src/plugin/spotters.rs
@@ -7,7 +7,7 @@ use yazi_shared::Id;
 use yazi_shim::arc_swap::IntoPointee;
 
 use super::Spotter;
-use crate::Selectable;
+use crate::plugin::SpotterMatcher;
 
 #[derive(Debug, Default, Deserialize)]
 pub struct Spotters(ArcSwap<Vec<Arc<Spotter>>>);
@@ -43,48 +43,5 @@ impl Spotters {
 
 	pub(crate) fn unwrap_unchecked(self) -> Vec<Arc<Spotter>> {
 		Arc::try_unwrap(self.0.into_inner()).expect("unique spotters arc")
-	}
-}
-
-// --- Matcher
-#[derive(Default)]
-pub struct SpotterMatcher<'a> {
-	pub spotters: Arc<Vec<Arc<Spotter>>>,
-	pub id:       Id,
-	pub file:     Option<Cow<'a, File>>,
-	pub mime:     Option<Cow<'a, str>>,
-	pub all:      bool,
-	pub offset:   usize,
-}
-
-impl From<&Spotters> for SpotterMatcher<'_> {
-	fn from(spotters: &Spotters) -> Self {
-		Self { spotters: spotters.load_full(), all: true, ..Default::default() }
-	}
-}
-
-impl SpotterMatcher<'_> {
-	pub fn matches(&self, spotter: &Spotter) -> bool {
-		if self.all {
-			true
-		} else if self.id != Id::ZERO {
-			spotter.id == self.id
-		} else {
-			spotter.match_with(self.file.as_deref(), self.mime.as_deref())
-		}
-	}
-}
-
-impl Iterator for SpotterMatcher<'_> {
-	type Item = Arc<Spotter>;
-
-	fn next(&mut self) -> Option<Self::Item> {
-		while let Some(spotter) = self.spotters.get(self.offset) {
-			self.offset += 1;
-			if self.matches(spotter) {
-				return Some(spotter.clone());
-			}
-		}
-		None
 	}
 }