+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 - New `--in` for `search` action to set search directory ([#3696])
 - Hover cursor over the new file after copying/cutting/linking/hardlinking/extracting ([#3846], [#3854])
 - Multi-file spotter ([#3733])
+- Dynamic open/opener Lua API ([#3901])
 - Dynamic previewer Lua API ([#3891])
 - Vim-like `lua` action that runs an inline Lua snippet ([#3813])
 - Certificate authentication for SFTP VFS provider ([#3716])
@@ -1703,3 +1704,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 [#3862]: https://github.com/sxyazi/yazi/pull/3862
 [#3891]: https://github.com/sxyazi/yazi/pull/3891
 [#3894]: https://github.com/sxyazi/yazi/pull/3894
+[#3901]: https://github.com/sxyazi/yazi/pull/3901